Refactor noScriptInfo.js with Aphrodite - 1st step
- Moved the file from js/components to app/renderer/components
- Moved styles of "allowScriptsButtons" from less/forms.less

Auditors:

Test Plan:
