+++ js/components/main.js
@@ -40,7 +40,7 @@ const ReleaseNotes = require('./releaseNotes')
 const BookmarksToolbar = require('../../app/renderer/components/bookmarksToolbar')
 const ContextMenu = require('./contextMenu')
 const PopupWindow = require('./popupWindow')
-const NoScriptInfo = require('./noScriptInfo')
+const NoScriptInfo = require('../../app/renderer/components/noScriptInfo')
 const LongPressButton = require('./longPressButton')
 const Menubar = require('../../app/renderer/components/menubar')
 const WindowCaptionButtons = require('../../app/renderer/components/windowCaptionButtons')
