+++ less/forms.less
@@ -515,9 +515,6 @@ select {
 
     // Issue #6783
     .allowScriptsButtons {
-      display: flex;
-      flex-flow: column nowrap;
-      align-items: center;
 
       > .browserButton {
         margin: 2px 0;
